Refactor: Change Docs button to a link
The Docs button was previously a button that opened a new window via a JavaScript method. This commit changes it to a standard anchor tag (`<a>`) that opens the documentation in a new tab using `target="_blank"`. This simplifies the code and provides a more standard user experience. The `openDocs` method is also removed as it is no longer needed.